requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
python-dotenv==0.10.1
requests==2.21.0
silverpeak==0.0.3.7
six==1.12.0
slackclient==1.3.1
tabulate==0.8.3
urllib3==1.24.3
websocket-client==0.54.0
